change database test location to exclude from action test
